fix: routes in all of the apps
